% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covert_m.R
\name{covert_m}
\alias{covert_m}
\title{converting unit(from hour to minute)}
\usage{
covert_m(data, x)
}
\arguments{
\item{data}{data.frame including jikan hun that is japanese charctor}

\item{x}{data.frame colm number including jikan hun}
}
\description{
converting unit(from hour to minute)
}
\examples{
\dontrun{
# converting "y hour yy minutes" to "zz minutes"
# Jpanese character only
time    <- c("1h20m","43m","20m","22m")
or <- c("kobe","osaka","kyoto","shiga")
dataF <- data.frame(origin=or,time=time)

# result
DD <- covert_m(dataF,2)
}

}
\author{
Keigo Matsuo \email{m.keigo.sp.ku@gmail.com}
}
