% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dest_loc.R
\name{dest_loc}
\alias{dest_loc}
\title{facility geocoding}
\usage{
dest_loc(data, x)
}
\arguments{
\item{data}{data.frame}

\item{x}{colm number colname must be "location"}
}
\description{
facility geocoding
}
\examples{
\dontrun{
# geocoding
# Jpanese character only
# colm name must be "location"
location <- c("Kobe university","Koshien studiam")
dataD <- data.frame(location=location)
dest.loc(dataD,1)
# result
#location latitude longitude
#Kobe University 34.72562  135.2354
#Kosien studiam 34.72129  135.3616
}

}
\author{
Keigo Matsuo \email{m.keigo.sp.ku@gmail.com}
}
