% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit.R
\name{transit}
\alias{transit}
\title{Get the transit data}
\usage{
transit(origin, destination, st_timeh, st_time2, st_time1)
}
\arguments{
\item{origin}{character of place name}

\item{destination}{character of place name}

\item{st_timeh}{time houre numeric}

\item{st_time2}{time minutes numeric}

\item{st_time1}{time minutes numeric}
}
\description{
Get the transit data
}
\examples{
\dontrun{
# from Kobe university to Osaka station departure:12:20
# Jpanese character only
transit(Kobe university, Osaka station, 12, 2, 0)
}

}
\author{
Keigo Matsuo \email{m.keigo.sp.ku@gmail.com}
}
