% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit_map_fare.R
\name{transit_map_dura}
\alias{transit_map_dura}
\title{mapping fare}
\usage{
transit_map_dura(orlon, orlat, zoom, data)
}
\arguments{
\item{orlon}{map center longitude}

\item{orlat}{map center latitude}

\item{zoom}{ggmap zoom}

\item{data}{data.frame including longitude,latitude,fare}
}
\description{
mapping fare
}
\examples{
\dontrun{
# from Kobe university to Osaka station departure:12:20
# center:Kobe University
transit.map.dura(135.235395,34.725918,11,Data)
}

}
\author{
Keigo Matsuo \email{m.keigo.sp.ku@gmail.com}
}
